backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
anthropic>=0.40.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.20.0
tavily-python>=0.3.0
